assistant.py
import requests
import os

# Assuming you have already loaded your environment variables (API keys, etc.)
from dotenv import load_dotenv

# Load environment variables from the .env file
load_dotenv()

# Define the Google Gemini API endpoint
GEMINI_API_URL = "https://api.google.com/gemini"  # Replace with actual Gemini API URL
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")  # Your Gemini API Key

def generate_financial_reply(question: str, as_html=False):
    """
    Generate a financial reply from the Google Gemini API or any other method.
    Args:
        question (str): The user's financial question.
        as_html (bool): If True, returns the response in HTML format. Default is False.

    Returns:
        str: The response from the financial assistant (Gemini).
    """
    # Send the question to Gemini API and get the response (this part can be customized)
    try:
        # Replace with your Gemini API request logic
        response = requests.post(
            GEMINI_API_URL,
            headers={"Authorization": f"Bearer {GEMINI_API_KEY}"},
            json={"query": question}
        )

        if response.status_code == 200:
            # Extract the relevant information from the Gemini API response
            response_data = response.json()

            # Example of how the response might look (you can adjust it based on the actual API response format)
            answer = response_data.get("response", "Sorry, I couldn't find an answer to your question.")

            # If `as_html` is True, return a simple HTML formatted response
            if as_html:
                return f"<p>{answer}</p>"

            return answer

        else:
            return "Sorry, there was an error while fetching the response from Gemini."

    except Exception as e:
        return f"Error: {e}"


def extract_text_from_file(file_bytes: bytes, filename: str) -> str:
    """
    Extract text from a file (PDF, CSV, Image, etc.) using a custom parser or an external library.
    Args:
        file_bytes (bytes): The file's byte content.
        filename (str): The filename of the uploaded document.

    Returns:
        str: The extracted text content.
    """
    # Dummy function to simulate text extraction (replace with actual logic)
    if filename.endswith(".pdf"):
        return extract_text_from_pdf(file_bytes)
    elif filename.endswith(".csv"):
        return extract_text_from_csv(file_bytes)
    elif filename.endswith((".jpg", ".png", ".jpeg")):
        return extract_text_from_image(file_bytes)
    else:
        return "Unable to extract content from the file."


def extract_text_from_pdf(file_bytes: bytes) -> str:
    """
    Extract text from a PDF file.
    Args:
        file_bytes (bytes): The byte content of the PDF.

    Returns:
        str: The extracted text from the PDF.
    """
    # Example: Use a PDF parser like PyMuPDF, PyPDF2, or pdfminer.six
    try:
        # Dummy implementation, replace with actual PDF text extraction logic
        return "Extracted PDF text content."
    except Exception as e:
        return f"Error extracting text from PDF: {e}"


def extract_text_from_csv(file_bytes: bytes) -> str:
    """
    Extract text from a CSV file.
    Args:
        file_bytes (bytes): The byte content of the CSV.

    Returns:
        str: The extracted text from the CSV.
    """
    # Example: Use the `csv` module to extract data from CSV
    try:
        # Dummy implementation, replace with actual CSV text extraction logic
        return "Extracted CSV content."
    except Exception as e:
        return f"Error extracting text from CSV: {e}"


def extract_text_from_image(file_bytes: bytes) -> str:
    """
    Extract text from an image file using OCR (Optical Character Recognition).
    Args:
        file_bytes (bytes): The byte content of the image.

    Returns:
        str: The extracted text from the image.
    """
    # Example: Use an OCR library like Tesseract to extract text from images
    try:
        # Dummy implementation, replace with actual image text extraction logic
        return "Extracted image text content."
    except Exception as e:
        return f"Error extracting text from image: {e}"
