requirements.txt
Flask==3.0.3
python-telegram-bot==20.1
PyPDF2==1.26.0
pytesseract==0.3.9
pillow==8.2.0
pandas==1.4.2
requests==2.27.1
python-dotenv==0.19.2
gunicorn==20.1.0  # Required for deploying to Heroku


